refactor js converters for shared use
